Split out ResultServer
